From: adrian@waltham.harvard.net Subject: [HACKERS] Another patch to configure.in
I heard very little in objections/approvals to defaulting some of the
parameters to configure.  Enclosed is a patch to configure.in which
removes the questions for

        PGPORT
        USE_LOCALE
        NOHBA

By default (i.e. assuming you don't put anything extra in the configure
command line), it assumes PGPORT=5432, USE_LOCAL=no and NOHBA=no (i.e.
HBA is turned on)

        --with-pgport=PGPORT_NO         Over-rides the PGPORT value
        --enable-locale                 enables USE_LOCALE
        --disable-hba                   disables HBA

Just for completeness:

        --prefix=BASEDIR                Defaults to /usr/local/pgsql
        --with-template=TEMPLATE        Defaults to asking you