Here are 3 patches (all relative to the src directory) to help with the configuration of v6.3.1.  I have replaced the queries for include/lib directories with --with configuration options.  I have also included a list of potential tcl/tk include directories directly in the CPPFLAGS variable.  As new versions are needed, these should be added to the list in reverse numerical order (libraries are in a separate list near the end).  This greatly simplifies the later checks if --with-tcl is set.  I hope this solution works for everyone.
I also added a check to disable the perl support if postgres was
not already installed (as per the instructions in the directory).
By the way, why must there be an installed pgsql to compile perl
support? This seems odd, at best.

Finally, I changed the Makefile in the libpgtcl interface to place
the shared libraries at the end of the list of files, not at the
beginning.  With NetBSD at least, libraries are linked in order,
so the original sequence does not work.

Brook Milligan