The check of class string in configure (line 1732) fails because this class is in <string> and not in <string.h> on QNX4/egcs-2.91.60. Probably this can be changed for all platforms. The test in line 1705 uses <string> as well. Because I am not sure, I havn't this included into the patch.
doc/Makefile has to be sligthly  modified as it has been done for
src/backend/Makefile due to a QNX4 problem (patch attached)

Furthermore src/test/regress/run_check.sh needs to be patched as it has been
done for regress.sh (patch attached). Please note that in the patch the
postmaster is started always with the -i option.

run_check.sh reports the test "limit" as failed, but in reallity it is OK.
regress.sh reports it as OK.

Andreas Kardos