Here is a bug and patch to fix it. I have tested this bug and fix on FreeBSD/Intel and DecUX/Alpha machines. The bug appears in postgresql 6.5.3 and 7.0.2. Can someone please review it and apply it to the source tree?
Sometimes when the postgres connection dies it is necessary to
attempt to reconnect. Calling the pgconnection::Connect method in a
derived class leaks memory because it does not clear the current
connection (if there is one). These patches ensures that any open
connections are closed before attempting to open a new one.

-Michael Richards