Incrementing version number in preparation for next release.  Note that I am talking with Thomas Lockhart about the idea of bringing the PyGreSQL version number into alignment with PostgreSQL so this may change to 7.1 before the release.
I have added to the copyright to indicate that from now on the PostgreSQL
copyright will apply.  If someone wants to make that clearer please do.
The existing copyrights need to stay there for now but if necessary I can
ask Pascal Andre if he agrees to a different wording.

Added reference to the Python DB-API 2.0 compliant API wrapper.

Added reference to the PyGreSQL mailing list.