Rewrite of planner statistics-gathering code.  ANALYZE is now available as a separate statement (though it can still be invoked as part of VACUUM, too). pg_statistic redesigned to be more flexible about what statistics are stored.  ANALYZE now collects a list of several of the most common values, not just one, plus a histogram (not just the min and max values).  Random sampling is used to make the process reasonably fast even on very large tables.  The number of values and histogram bins collected is now user-settable via an ALTER TABLE command.
There is more still to do; the new stats are not being used everywhere
they could be in the planner.  But the remaining changes for this project
should be localized, and the behavior is already better than before.

A not-very-related change is that sorting now makes use of btree comparison
routines if it can find one, rather than invoking '<' twice.