drop_function.sgml: =================== In Notes:  Refer to CREATE FUNCTION for information on creating aggregate functions.                                                       ^^^^^^^^^^^^^^^^^^^ I assume it must read C function instead.
In Compatibility SQL/PSM:
 SQL/PSM is a proposed standard.

 We had that before: remove proposed.


drop_index.sgml:
================
<REFNAME>: Removes existing indexes from a database

as far as I can see index should be singular. The command description is
written as if only one index can be removed at a time. Interestingly
enough, in v7.0.2 it was in fact singular. Am I mistaken here?

drop_operator.sgml:
===================
In Outputs the arguments are referred to as type and type2, but the synopsis
and Inputs section these are left_type and right_type, respectively. Also,
oper is used in Outputs versus id in Inputs/Synopsis. In the translation I
follow the replaceables used in the Inputs/Synopsis part.

Frank Wegmann