Attached is my patch that adds DROP CONSTRAINT support to PostgreSQL.  I basically want your guys feedback.  I have sprinkled some of my q's thru the text delimited with the @@ symbol.  It seems to work perfectly.
[ Removed @@ comments because patch was reviewed. ]

At the moment it does CHECK constraints only, with inheritance.  However,
due to the problem mentioned before with the mismatching between inherited
constraints it may be wise to disable the inheritance feature for a while.
it is written in an extensible fashion to support future dropping of other
types of constraint, and is well documented.

Please send me your comments, check my use of locking, updating of
indices, use of ERROR and NOTICE, etc. and I will rework the patch based
on feedback until everyone
is happy with it...

Christopher Kings