> Huh, I don't know where I got the idea you were (or someone else was?) > in the position that attislocal should be reset.  I'll clean everything > up and submit the patch I had originally made.
All right, this is it.  This patch merely checks if child tables have
the column.  If atttypid and atttypmod are the same, the attributes'
attinhcount is incremented; else the operation is aborted.  If child
tables don't have the column, recursively add it.

attislocal is not touched in any case.

Alvaro Herrera